+++ modules/canbus/vehicle/lexus/protocol/steering_cmd_12c.cc
@@ -64,7 +64,7 @@ Steeringcmd12c* Steeringcmd12c::set_ignore_overrides(bool ignore_overrides) {
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Steeringcmd12c::set_p_ignore_overrides(uint8_t* data,
                                             bool ignore_overrides) {
-  int x = ignore_overrides;
+  uint8_t x = ignore_overrides;
 
   Byte to_set(data + 0);
   to_set.set_value(static_cast<uint8_t>(x), 1, 1);
@@ -79,7 +79,7 @@ Steeringcmd12c* Steeringcmd12c::set_enable(bool enable) {
 // 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 0, 'type': 'bool',
 // 'order': 'motorola', 'physical_unit': ''}
 void Steeringcmd12c::set_p_enable(uint8_t* data, bool enable) {
-  int x = enable;
+  uint8_t x = enable;
 
   Byte to_set(data + 0);
   to_set.set_value(static_cast<uint8_t>(x), 0, 1);
@@ -94,7 +94,7 @@ Steeringcmd12c* Steeringcmd12c::set_clear_override(bool clear_override) {
 // 'len': 1, 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 2,
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Steeringcmd12c::set_p_clear_override(uint8_t* data, bool clear_override) {
-  int x = clear_override;
+  uint8_t x = clear_override;
 
   Byte to_set(data + 0);
   to_set.set_value(static_cast<uint8_t>(x), 2, 1);
@@ -109,7 +109,7 @@ Steeringcmd12c* Steeringcmd12c::set_clear_faults(bool clear_faults) {
 // 'len': 1, 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 3,
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Steeringcmd12c::set_p_clear_faults(uint8_t* data, bool clear_faults) {
-  int x = clear_faults;
+  uint8_t x = clear_faults;
 
   Byte to_set(data + 0);
   to_set.set_value(static_cast<uint8_t>(x), 3, 1);
